Fixed autocompletion
Since one of the recent updates broke the autocompletion function due to changing class names, I took the liberty to attempt to fix it. Since this is just a quick fix I threw together some classes may be wrong and need further correction. The functionality of the autocompletion function was however restored.