Adjust match check for BBD "Normalize Classes"
BandagedBD has a "Normalize Classes" feature I forgot is enabled and adds classes to a bunch of things.  This fixes the check and should make it more flexible to handle if extra classes are added down the road without breaking the plugin.  Based on what I saw the hljs class is only used with code blocks and this looks to see if it is ending with that style.