Fix LineNumbers for latest Discord
Trying to make it future proof.  Only checks for code block styles ending with hljs instead of trying to match a whole style run.  This also allows the plugin to work more nicely with BandagedBD's "Normalize Styles" function since it inserts in additional styles in the middle causing a match fail but still ends with hljs with that as well.